+++ packages/dashboard/src/components/Card/Layout/Layout.tsx
@@ -218,7 +218,14 @@ export class Layout extends React.Component<ILayoutProps, { _width: number; _hei
     if (header === null || header === undefined) {
       return null;
     }
-    return <CardHeader headerText={header.headerText} annotationText={header.annotationText} fontSize={header.fontSize} />;
+    return (
+      <CardHeader
+        headerText={header.headerText}
+        annotationText={header.annotationText}
+        fontSize={header.fontSize}
+        cardHeaderCallback={header.cardHeaderCallback}
+      />
+    );
   }
 
   private _generateFooter(