+++ packages/foundation/src/createComponent.tsx
@@ -79,6 +79,8 @@ export function createComponent<
 
             const viewComponentProps: typeof mergedProps & IDefaultSlotProps<any> = {
               ...(mergedProps as any),
+              styles,
+              tokens,
               _defaultStyles: styles
             };