+++ packages/office-ui-fabric-react/src/components/Callout/CalloutContent.base.tsx
@@ -100,6 +100,14 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
     }
   }
 
+  public shouldComponentUpdate(newProps: ICalloutProps): boolean {
+    if (this.props.hidden && newProps.hidden) {
+      // Do not update when hidden.
+      return false;
+    }
+    return true;
+  }
+
   public componentWillMount() {
     this._setTargetWindowAndElement(this._getTarget());
   }
@@ -121,8 +129,11 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
       this._setHeightOffsetEveryFrame();
     }
 
-    // if the callout becomes hidden, then remove any positions, bounds that were placed on it.
-    if (newProps.hidden && newProps.hidden !== this.props.hidden) {
+    // Ensure positioning is recalculated when we are about to show a persisted menu.
+    if (!newProps.hidden && newProps.hidden !== this.props.hidden) {
+      this._maxHeight = undefined;
+      // Target might have been updated while hidden.
+      this._setTargetWindowAndElement(newTarget);
       this.setState({
         positions: undefined
       });