+++ packages/office-ui-fabric-react/src/components/Dropdown/examples/Dropdown.Error.Example.tsx
@@ -1,20 +1,27 @@
 import * as React from 'react';
 import { Dropdown } from 'office-ui-fabric-react/lib/Dropdown';
-import { BaseComponent } from 'office-ui-fabric-react/lib/Utilities';
-import './Dropdown.Basic.Example.scss';
+import { IStackTokens, Stack } from 'office-ui-fabric-react/lib/Stack';
+import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
 
-export class DropdownErrorExample extends BaseComponent<{}, {}> {
-  constructor(props: {}) {
-    super(props);
-  }
+export interface IDropdownErrorExampleState {
+  showError: boolean;
+}
 
-  public render(): JSX.Element {
+export class DropdownErrorExample extends React.Component<{}, IDropdownErrorExampleState> {
+  // Don't show the error message by default because it's annoying to screen reader users.
+  public state: IDropdownErrorExampleState = { showError: false };
+
+  public render() {
+    const { showError } = this.state;
+
+    const stackTokens: IStackTokens = { childrenGap: 30 };
+
     return (
-      <div className="docs-DropdownExample">
+      <Stack horizontal tokens={stackTokens} verticalAlign="start">
+        <Toggle label="Show error message" onText="Yes" offText="No" checked={showError} onChange={this._updateShowError} />
         <Dropdown
-          placeholder="Select an Option"
-          label="Error message example:"
-          ariaLabel="Error message dropdown example"
+          placeholder="Select an option"
+          label="Dropdown with error message"
           options={[
             { key: 'A', text: 'Option a' },
             { key: 'B', text: 'Option b' },
@@ -22,9 +29,14 @@ export class DropdownErrorExample extends BaseComponent<{}, {}> {
             { key: 'D', text: 'Option d' },
             { key: 'E', text: 'Option e' }
           ]}
-          errorMessage="Error message"
+          errorMessage={showError ? 'This dropdown has an error' : undefined}
+          styles={{ dropdown: { width: 300 }, root: { height: 100 } }}
         />
-      </div>
+      </Stack>
     );
   }
+
+  private _updateShowError = (event: React.MouseEvent<HTMLElement>, checked?: boolean) => {
+    this.setState({ showError: !!checked });
+  };
 }