+++ packages/react-cards/src/components/Card/Card.ts
@@ -1,16 +1,16 @@
+import { IStackItemProps } from 'office-ui-fabric-react';
+import { createComponent } from '@uifabric/foundation';
 import { CardView as view } from './Card.view';
 import { CardStyles as styles, CardTokens as tokens } from './Card.styles';
 import { ICardProps } from './Card.types';
 import { CardItem } from './CardItem/CardItem';
-import { ICardItemProps } from './CardItem/CardItem.types';
-import { createComponent } from '@uifabric/foundation';
 
 const CardStatics = {
   Item: CardItem
 };
 
 export const Card: React.FunctionComponent<ICardProps> & {
-  Item: React.FunctionComponent<ICardItemProps>;
+  Item: React.FunctionComponent<IStackItemProps>;
 } = createComponent({
   displayName: 'Card',
   view,