+++ apps/fabric-website-resources/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @uifabric/fabric-website-resources
 
-This log was last generated on Tue, 05 Mar 2019 04:25:07 GMT and should not be manually modified.
+This log was last generated on Mon, 01 Apr 2019 12:37:03 GMT and should not be manually modified.
+
+## 6.14.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- add Separator page
+
+## 6.13.2
+Fri, 29 Mar 2019 12:36:45 GMT
+
+### Patches
+
+- DemoPage: pass all extra props to ComponentPage
+
+## 6.13.1
+Wed, 27 Mar 2019 12:34:02 GMT
+
+### Patches
+
+- Improve imports of color utilities
 
 ## 6.13.0
 Tue, 05 Mar 2019 04:25:07 GMT