+++ packages/react-cards/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/react-cards
 
-This log was last generated on Tue, 26 Mar 2019 12:32:51 GMT and should not be manually modified.
+This log was last generated on Mon, 01 Apr 2019 12:37:03 GMT and should not be manually modified.
+
+## 0.2.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- Card: Improving styling and adding compact Card styling and examples.
+
+## 0.1.2
+Wed, 27 Mar 2019 12:34:02 GMT
+
+### Patches
+
+- remove old api file
 
 ## 0.1.1
 Tue, 26 Mar 2019 12:32:51 GMT